Add unary operator minus
