+++ calc1.c
@@ -18,91 +18,99 @@ typedef struct Interpreter
     Token current_token;
 } Interpreter;
 
+int factor(Interpreter *);
+int term(Interpreter *);
+void expr(Interpreter *);
 void interpret(Interpreter *);
 Token get_next_token(Interpreter *);
 void parse (Interpreter *,char *);
 void parse_error();
 
+int factor(Interpreter *i)
+{
+	int curfactor;
+	Token fact = i->current_token;
+	
+	if (strcmp(i->current_token.type,"MINUS") == 0)
+	{
+		parse(i,"MINUS");
+		
+		fact = i->current_token;
+		parse(i,"INTEGER");
+		
+		return (-1) * atoi(fact.value.ptr);
+	}
+	else
+		parse(i,"INTEGER");
+	
+	curfactor = atoi(fact.value.ptr);
+	return curfactor;
+}
+
+int term(Interpreter *i)
+{
+	int t = factor(i);
+	if (strcmp(i->current_token.type,"INTEGER") == 0)
+		parse_error();
+	while ((strcmp(i->current_token.type,"MULTIPLY") == 0) || (strcmp(i->current_token.type,"DIVIDE") == 0))
+	{
+		Token tok = i->current_token;
+		if (strcmp(tok.type,"MULTIPLY") == 0)
+		{
+			parse(i,"MULTIPLY");
+			t = t * factor(i);
+		}
+		else
+		{
+			parse(i,"DIVIDE");
+			t = t / factor(i);
+		}
+	}
+	return t;
+}
+
+void expr(Interpreter *i)
+{
+	//Token result;
+	int res = term(i);
+	while ((strcmp(i->current_token.type,"PLUS") == 0) || (strcmp(i->current_token.type,"MINUS") == 0) || (strcmp(i->current_token.type,"MODULUS") == 0))
+	{
+		Token tok = i->current_token;
+		if (strcmp(tok.type,"PLUS") == 0)
+		{
+			parse(i,"PLUS");
+			res = res + term(i);
+		}
+		else if (strcmp(tok.type,"MINUS") == 0)
+		{
+			parse(i,"MINUS");
+			res = res - term(i);
+		}
+		else
+		{
+			parse(i,"MODULUS");
+			res = res % term(i);
+		}
+	}
+	//result.type = "INTEGER";
+	//result.value.ptr = "10";
+	//return result;
+	printf("%d\n",res);
+}
+
 void interpret(Interpreter *i)
 {
-    Token left,op,right;
-    int result,flag = 0;
-
-    i->pos = 0;
-    i->current_token.type = "";
-    init_string(&i->current_token.value,MAXLEN);
-
-    i->current_token = get_next_token(i);
-    left = i->current_token;
-    parse(i,"INTEGER");
-    //printf("%s %d %s %s\n",i->text,i->pos,i->current_token.type,i->current_token.value.ptr);
-    while(strcmp(i->current_token.type,"EOF") != 0)
-    {
-        op = i->current_token;
-
-        if(strcmp(op.type,"PLUS") == 0)
-        {
-            parse(i,"PLUS");
-            flag = 1;
-        }
-        else if(strcmp(op.type,"MINUS") == 0)
-        {
-            parse(i,"MINUS");
-            flag = 2;
-        }
-        else if(strcmp(op.type,"MULTIPLY") == 0)
-        {
-            parse(i,"MULTIPLY");
-            flag = 3;
-        }
-        else if(strcmp(op.type,"DIVIDE") == 0)
-        {
-            parse(i,"DIVIDE");
-            flag = 4;
-        }
-	else if(strcmp(op.type,"MODULUS") == 0)
-	{
-		parse(i,"MODULUS");
-		flag = 5;
-	}
-        else
-            parse_error();
-
-        right = i->current_token;
-        parse(i,"INTEGER");
-
-        switch(flag)
-        {
-        case 1:
-            result = atoi(left.value.ptr) + atoi(right.value.ptr);
-            break;
-        case 2:
-            result = atoi(left.value.ptr) - atoi(right.value.ptr);
-            break;
-        case 3:
-            result = atoi(left.value.ptr) * atoi(right.value.ptr);
-            break;
-        case 4:
-            result = atoi(left.value.ptr) / atoi(right.value.ptr);
-            break;
-	case 5:
-	    result = atoi(left.value.ptr) % atoi(right.value.ptr);
-	    break;
-        }
-
-        snprintf(left.value.ptr, sizeof(left.value.ptr),"%d", result);
-    }
-    if(flag)
-        printf("%d\n",result);
-    else
-        display_string(left.value);
-
-    free(i->current_token.value.ptr);
+	//space check from main interpret function
+	i->pos = 0;
+	i->current_token.type = "";
+	init_string(&i->current_token.value,MAXLEN);
+	i->current_token = get_next_token(i);
+	expr(i);
 }
+	
 
 void parse(Interpreter *i,char *type)
 {
-    //printf("%s %d %s %s\n",i->text,i->pos,i->current_token.type,i->current_token.value.ptr);
     if(strcmp(i->current_token.type,type) == 0)
         i->current_token = get_next_token(i);
     else
